# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel7CRUD OpenApi Documentation",
        "description": "Based on L5 Swagger package",
        "contact": {
            "email": "karasickdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login registered user",
                "description": "Returns authenticated user data",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentialsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthUserDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationException"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout authenticated post",
                "description": "Disable authenticated user token and returns no content",
                "operationId": "logoutUser",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Autorization Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 30|TPKtQDvDzhyNgdms51MAta6HJcM6CXQwgrRcYxFp"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "description": "Returns authenticated user data",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDataRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthUserDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationException"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/checkAccessToken": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check registered user access token",
                "description": "Returns authenticated user data",
                "operationId": "checkUserAccessToken",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Autorization Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 30|TPKtQDvDzhyNgdms51MAta6HJcM6CXQwgrRcYxFp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthUserDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BadRequestException": {
                "title": "Bad Request Exception",
                "description": "Bad Request Exception response data",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Validation Exception message",
                        "type": "string",
                        "example": "Data cannot be validated."
                    }
                },
                "type": "object"
            },
            "ValidationException": {
                "title": "Validation Exception",
                "description": "Validation Exception response data",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Validation Exception message",
                        "type": "string",
                        "example": "Data cannot be validated."
                    },
                    "errors": {
                        "title": "errors",
                        "description": "Validation Exception errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "AuthUserDto": {
                "title": "AuthUser",
                "description": "Authenticated User DTO",
                "properties": {
                    "token": {
                        "title": "token",
                        "description": "User Access Token",
                        "type": "string",
                        "example": "token"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDto"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthUserDto"
                }
            },
            "UserDto": {
                "title": "User",
                "description": "User DTO",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "johndoe@mail.com"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserDto"
                }
            },
            "UserCredentialsRequest": {
                "title": "User Credentials Request",
                "description": "User Credentials request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of the registered user",
                        "type": "string",
                        "example": "johndoe@mail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the registered user",
                        "type": "string",
                        "example": "password1"
                    }
                },
                "type": "object"
            },
            "UserDataRequest": {
                "title": "User Data Request",
                "description": "User Data request body data",
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "johndoe@mail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password1"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "Password confirmation of the user",
                        "type": "string",
                        "example": "password1"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}